requirements.txt
beautifulsoup4 
sqlalchemy 
psycopg2-binary 
flask 
apscheduler 
dash 
pandas
streamlit
requests
python-dotenv